functions/requirements.txt
firebase-admin==6.4.0
google-cloud-firestore==2.13.1
pydantic==2.5.2
firebase_functions==0.1.0
functions-framework